+++ programmes/forms.py
@@ -17,201 +17,6 @@ from programmes.models import (
 )
 
 
-class ProgrammeSelectionForm(forms.Form):
-    def __init__(
-        self, *args, lots=None, bailleurs=None, administrations=None, **kwargs
-    ) -> None:
-        self.declared_fields["lot"].choices = lots
-        self.declared_fields["bailleur"].choices = bailleurs
-        self.declared_fields["administration"].choices = administrations
-        super().__init__(*args, **kwargs)
-
-    lot = forms.CharField(
-        required=False,
-        label="Lot à conventionner",
-        error_messages={
-            "required": "La selection du programme et de son financement est obligatoire"
-        },
-    )
-    existing_programme = forms.ChoiceField(
-        choices=[("selection", "selection"), ("creation", "creation")]
-    )
-    bailleur = forms.ChoiceField(
-        required=False,
-        label="Bailleur",
-        choices=[],
-        error_messages={
-            "required": "Le bailleur est obligatoire",
-        },
-    )
-    administration = forms.ChoiceField(
-        required=False,
-        label="Administration",
-        choices=[],
-        help_text="Délégataire des aides à la pierre du territoire de l'opération",
-        error_messages={
-            "required": "L'administration est obligatoire",
-        },
-    )
-    nom = forms.CharField(
-        required=False,
-        label="Nom du programme",
-        max_length=255,
-        min_length=1,
-        error_messages={
-            "required": "Le nom du programme est obligatoire",
-            "min_length": "Le nom du programme est obligatoire",
-            "max_length": "Le nom du programme ne doit pas excéder 255 caractères",
-        },
-    )
-    nb_logements = forms.IntegerField(
-        required=False,
-        label="Nb logements à conventionner",
-        error_messages={
-            "required": "Le nombre de logements à conventionner est obligatoire",
-        },
-    )
-    type_habitat = forms.TypedChoiceField(
-        required=False,
-        label="Type d'habitat",
-        choices=TypeHabitat.choices,
-        error_messages={
-            "required": "Le type d'habitat est obligatoire",
-        },
-    )
-    financement = forms.TypedChoiceField(
-        required=False,
-        label="Financement",
-        choices=FinancementEDD.choices,
-        error_messages={
-            "required": "Le financement est obligatoire",
-        },
-    )
-    code_postal = forms.CharField(
-        required=False,
-        label="Code postal",
-        max_length=255,
-        error_messages={
-            "required": "Le code postal est obligatoire",
-            "max_length": "Le code postal ne doit pas excéder 255 caractères",
-        },
-    )
-    ville = forms.CharField(
-        required=False,
-        label="Ville",
-        max_length=255,
-        error_messages={
-            "required": "La ville est obligatoire",
-            "max_length": "La ville ne doit pas excéder 255 caractères",
-        },
-    )
-
-    def validate_required_field(self, cleaned_data, field_name):
-        if field_name in cleaned_data and (
-            cleaned_data[field_name] is None or cleaned_data[field_name] == ""
-        ):
-            self._errors[field_name] = self.error_class(
-                [self.fields[field_name].error_messages["required"]]
-            )
-            del cleaned_data[field_name]
-
-    def clean(self):
-        cleaned_data = super().clean()
-        if cleaned_data["existing_programme"] == "selection":
-            self.validate_required_field(cleaned_data, "lot_uuid")
-        if cleaned_data["existing_programme"] == "creation":
-            self.validate_required_field(cleaned_data, "bailleur")
-            self.validate_required_field(cleaned_data, "administration")
-            self.validate_required_field(cleaned_data, "nom")
-            self.validate_required_field(cleaned_data, "nb_logements")
-            self.validate_required_field(cleaned_data, "financement")
-            self.validate_required_field(cleaned_data, "type_habitat")
-            self.validate_required_field(cleaned_data, "code_postal")
-            self.validate_required_field(cleaned_data, "ville")
-
-
-class ProgrammeSelectionFromDBForm(forms.Form):
-    def __init__(self, *args, lots=None, **kwargs) -> None:
-        self.declared_fields["lot"].choices = lots
-        super().__init__(*args, **kwargs)
-
-    lot = forms.CharField(
-        label="Lot à conventionner",
-        error_messages={
-            "required": "La selection du programme et de son financement est obligatoire"
-        },
-    )
-
-
-class ProgrammeSelectionFromZeroForm(forms.Form):
-    def __init__(self, *args, bailleurs=None, administrations=None, **kwargs) -> None:
-        self.declared_fields["bailleur"].choices = bailleurs
-        self.declared_fields["administration"].choices = administrations
-        super().__init__(*args, **kwargs)
-
-    bailleur = forms.ChoiceField(
-        label="Bailleur",
-        choices=[],
-        error_messages={
-            "required": "Le bailleur est obligatoire",
-        },
-    )
-    administration = forms.ChoiceField(
-        label="Administration",
-        choices=[],
-        help_text="Délégataire des aides à la pierre du territoire de l'opération",
-        error_messages={
-            "required": "L'administration est obligatoire",
-        },
-    )
-    nom = forms.CharField(
-        label="Nom du programme",
-        max_length=255,
-        min_length=1,
-        error_messages={
-            "required": "Le nom du programme est obligatoire",
-            "min_length": "Le nom du programme est obligatoire",
-            "max_length": "Le nom du programme ne doit pas excéder 255 caractères",
-        },
-    )
-    nb_logements = forms.IntegerField(
-        label="Nb logements à conventionner",
-        error_messages={
-            "required": "Le nombre de logements à conventionner est obligatoire",
-        },
-    )
-    type_habitat = forms.TypedChoiceField(
-        label="Type d'habitat",
-        choices=TypeHabitat.choices,
-        error_messages={
-            "required": "Le type d'habitat est obligatoire",
-        },
-    )
-    financement = forms.TypedChoiceField(
-        label="Financement",
-        choices=FinancementEDD.choices,
-        error_messages={
-            "required": "Le financement est obligatoire",
-        },
-    )
-    code_postal = forms.CharField(
-        label="Code postal",
-        max_length=255,
-        error_messages={
-            "required": "Le code postal est obligatoire",
-            "max_length": "Le code postal ne doit pas excéder 255 caractères",
-        },
-    )
-    ville = forms.CharField(
-        label="Ville",
-        max_length=255,
-        error_messages={
-            "required": "La ville est obligatoire",
-            "max_length": "La ville ne doit pas excéder 255 caractères",
-        },
-    )
-
-
 class ProgrammeForm(forms.Form):
     object_name = "programme"
